Only send map config commands at the beginning
Sometimes map config commands could be sent at the end of the old map,
instead of the beginning of the new map. This made no difference because
the commands are sent *after* the map command. So the commands execute
after the first sgame frame in either case. Only send commands at the
beginning because it's simpler.