Exec map configs before map_restart when possible
Before, the map config commands were sometimes sent at the previous map
and sometimes at the beginning of the new map, but always executed right
after map_restart (so there was no difference between those scenarios).

Now, map configs execute at the end of the previous map, before
map_restart,  for any map
change that originates from the sgame. (/map and /devmap are server
commands so it is impossible with those.) This could be useful to set
fs_extrapaks in a map config.