Generate navmeshes during cgame loading screen
This feature can be turned on by cg_navgenOnLoad. It only runs if the
user is the server host.

The loading progress bar restarts from 0 if navgen is needed.

It's done in the cgame instead of the sgame because it takes a long time
and you can't have a loading screen when the sgame is loading.