Improve blocker handling
This improves on #1493 by allowing bots to move on flat or on downward
tight corridors.
The code is a quick workaround, as explained in the commit, a correct
fix would require a complete rewrite of the obstacle management code,
and would even require better engine API to give bots a usable "vision".