RMLUI: fix cursor disappearing on vid_restart
In particular it happened when the console is closed on the main menu.

To prevent the bug from recurring in another case, when the console is
open and the mouse is outside the window after the vid_restart, a commit
to remove the duplicate call to IN_Init in Daemon is needed.