bg_pmove: one-linerize trace calls
This may make things harder to read for now, but it makes it easier to
think about how to find a way to refactor so that there would be less
ducplicate in code and more obvious semantics.
Currently, I think there may be a way to merge the VectorMA() call right
before in the function, and maybe to reduce the number of parameter to
get something easier to understand when reading code.
This file is still a rather critical file, so I'm doing very incremental
changes here.