debian: update packaging, also rewrite unvanquished script
The unvanquished-helper script is now renamed to unvanquished
naming it (or naming a symlink to it) to unvanquished-server
or unvanquished-tty will launch daemonded or daemon-tty instead.
It now ignores unvanquished-migrate-profile script if not there.

The dpk archives are now provided within unvanquished-data package
instead of providing a download script with (now removed)
unvanquished-data-downloader package.

The unvanquished .desktop file is updated and a separate protocol
handler one is added.