Map list: lazily load levelshots; ignore .arena
See https://github.com/DaemonEngine/Daemon/issues/557.

Before: When the map list is opened, all paks with the map- prefix are
opened and files with the meta/ prefix are opened. Arena files are
parsed. The map's name is displayed in the list according to the
'longname' key in the .arena.

After: The map list is built based on pak names that begin with "map-".
The meta/ prefix is loaded for a given pak only when you click on it in
the list. The pak name (excluding the prefix) is used as the name of the
map in the list. Arena files are not used.