Designate Quote() for escaping server commands
Instead of being just a proxy for Cmd::Escape, Quote now has a different
functionality to escape strings with newlines for server->client
commands. This fixes the bug that multiline outputs stop working after
Cmd::Escape was changed not to allow newlines.

Cases of Quote used for engine commands are changed to Cmd::Escape and
cases of Cmd::Escape used for server commands are changed to Quote.