fix trigger_class and trigger_equipement
Those triggers were broken compared to tremulous, where they could be
used notably to test for a team (if list is empty).
There is no clue at all that the non-deprecated trigger works, and no
clue that it is actually used, this patch only fixes stuff which is
used in practice.
Some actual examples of usage are:
* tremship (used to detect team, but also for some easter eggs I
  believe)
* unvanquished's 0.52 parpax (used to implement a joke "achievement" in
  vents)