Introduce fallback BT nodes
The name is taken from BT.cpp, https://docs.ros.org/en/kinetic/api/behaviortree_cpp_v3/html/classBT_1_1FallbackNode.html

The difference with `sequence` is that this one switches to the next one
if the first action *fails*. `sequence` goes to the next child when the
current child succeeds and stops when one action fails; `fallback` goes
to the next action when the current action fails and stops when one
action succeeded.

The difference with `selector` is that the selector falls tries the
first element of the list each time to select the first one which
doesn't fail, while `fallback` continues where it was on last
`ACTION_RUNNING`.