welcome: Fix showing welcome window on connection.
This used to be g_showHelpOnConnection, but that cvar is too long, so just call it g_showWelcome. Also, with the old code, it would come at the same frame that we go to CA_ACTIVE, which meant that rmlui would hide all the windows. However, since, PRIMED happens just before we get the first data from the server, it makes sense to hide all the menus then so we don't hide any commands we receive from the server.

Also, refactor the switch statement to be more readable instead of relying on wonky fallthroughs.