rmlui: Fix rocketFormControlSelect accidently setting cvars.
The old code naively just set the value of the form to the value of the cvar and hoped RmlUi would do something intelligent if the value set didn't mask any of the options. However, RmlUi will simply select the last value if it doesn't match any of the values. This is not what we want, so if be smarter by checking the values of all the options, if none of them match, dynamically add a new option for the current value.