rmlui: Remove the concept of ELEMENT_DEAD.
It started with good intentions of only showing certain elements when
players are dead. However, contrary to expectations, these enums control
when the Update and Render code for these functions is run. That means
that as soon as you spawn or become a spectator, the Update And Render
codepaths of these elements doesn't run and doesn't allow the element to
update itself to hide/change/etc when a player becomes not dead or a
spectator. In the respawn case, we were getting lucky because when you
respawn, you spawn with a weapon, which would show a new HUD. However,
when you become a spectator, you still have the same weapon as you did
when you were dead (WP_NONE), but now this code doesn't run making
element dead messages linger on spectator screens.

This fixes #2085