rmlui: Allow running commands from chat.
By prefixing your text with a /, you can run a command. This was the
behavior of the trem chat. Since you need to use commands to do a lot of
useful things in Unv, it's easy to use the chat to run these commands
instead of being forced to open the console.

This behavior can be disabled by using the ``noCmd="1"`` attribute in
the <chatField /> element.