chat: Show a chat bubble when players are typing.
This uses the usercmd flags which relate to input to note whether the
game has input. Currently, AFAICT this field is unused, so we use it to
transmit whether the client is typing. WE need to remove the RSF_SPRITE
from the chatBubble thing because for some reason the chat bubble
doesn't show with it.

Fixes #2217