Make new way to iterate entities with a component
To replace ForEntities since some people didn't like it. For each
component type there is, instead of a std::set with pointers to the
components, a MAX_GENTITIES-bit array indicating which entities have the
component. Instead of the loop calling a lambda you use a normal
for-each loop iterating over either the Entity's or the components.