rmlui: Bounds check the chat cursor.
We let you move the cursor forward every time you press right, even if
it is already at the end of the string, this can allow you to go past
the end of the string when you try to enter the next character which can
lead to a crash.

Thanks to @sweet235 for the report and repro instructions.

Fixes #2138