Remove unneeded destructor of SpawnerComponent
Detecting death in the destructor is unnecessary because the Die message
is always sent for a buildable that ceases to exist, even if it was
a "nonviolent" MOD like deconstructing.

If the code did run it would be a bug because OnLoss invokes the
TeamComponent and there is no guarantee that the TeamComponent isn't
destroyed before the SpawnerComponent.