Fix g_bot_defaultFill overriding bot fill
The bot fill target is set each time that either a `bot fill` command is
executed, or G_BotFill() checks the g_bot_defaultFill cvar and finds
that it was modified. The cvar is considered to be "modified" if it has
not been read in the current sgame execution. So ensure that the first
modification check occurs before reading map configs, so that if there
is a `bot fill` in map configs it overrides the value taken from the
cvar.