Remove command `bot debug_reload`
This has at least one known deficiency: the obstacle information
is lost, meaning bots will not be able to maneuver around any entities
that have existed since before the command was called. And this is
a particularly frustrating and hard to pin down bug, so debug_reload
is an attractive nuisance that is better off deleted.