Implement negative admin flags, which begin with a period [.] character. These will be listed in red in the /flaglist.
This allows a server admin to revoke permissions such as:
chatting (team and global)
voting (team and global)
building

on an individual and permanent basis using the /flag command, or by
having these flags set in the g_admin file if required.

This gives an additional element of granular control to server admins
and will aid moderation of the server.