Allow setting g_BPInitialBudget to a value per team
Setting "g_BPInitialBudget 10,50" will give 10 bp to the
first team (aliens) and 50bp to the second (humans).

If the list isn't long enough, the last value will be used. This means
that this is backward compatible as if there is only one value, it will
be used for every team.

This approach has the advantage of extending to more teams should we add
support for that; and doesn't need a change in netradiant for each team
we add either.