include part of the jump magnitude into STEPSIZE
The idea is simply to:

* let bots jump over non friendly stuff;
* generate a navmesh that **partially** includes the jump magnitude into
  stepsize;

The 2nd point can be dangerous, since too high values, relative to the
total height of the agent, can break pathfinding for reasons I still do
not know (linked to recast).
According to my past testing, a value of 0.625 is safe, but in case it
is not, the feature can be configured through the "autojump" key in
configuration system, including on a per-map basis.