Refactor: flatten the skill tree structure and make it more declarative
This moves away from predicates and nested structure to a flat vector of
skills that tells which skills they depend on, and which team they are
for.

This is a step towards better introspection of the tree and better
tooling.

This also allows having a skill behind either two skills, on choice.
Which makes that skill more easily accessible.

I've noticed that the bots get different skill sets for the same seed
after this change, I believe this is due to a change in the order of the
skills.