incognito behaviour changes
by popular decision on IRC: now admins can see <= their own level instead of needing the setlevel flag.
this commit introduces the SEESINCOGNITO flag which, when set, allows an admin to see all incognito admins regardless of their level.
this commit also fixes a leak in the "NOT AUTHED" indicator - which can now only be seen by admins with setlevel access and a higher admin level.