make the skillset feature opt-out
This works by adding a cvar named: g_bot_defaultSkill

If this cvar is empty, which is the default, behavior won't change.
Otherwise, it must be filled with "key:value" pairs, separated by a
space (' ').
When a bot is added that have a skill >= key, then it gets the "value"
skill automatically.
There are sanity checks that prevents non-existing skills to be tried to
be added.

In a reasonable futur, I'd like to add a "/bot listSkills" command that
will list all available skills. This could be then used to build a GUI
that helps building the cvar.