Implement Admin Stealth and Super Stealth.
This allows admins to be anonymous when using admin commands.
Stealthed admins will be identified to those with SEESSTEALTH flags
by way of a chat tag (which is explained in the adminhelp text).

This commit also increases the buffer for Quote( ) to accommodate this
change and fixes all known bugs with trailing $x$ in admin commands.