make automatic crouching opt-out
Previous commits only affected the navgen, but bots would still crouch
in some cases (i.e. a turret built in the way that need crouch or
steering) which is usually a desired behavior.
It has been reported that sometimes bots crouch for nothing though, so
this commit allows server owner to disable it on some maps, or enterily.