Fix intermittently disappearing mara zaps
Thanks to global variable abuse, `muzzle` was never calculated when
calling UpdateZapEffect via G_UpdateZaps. Each frame, the origin of the
zap entity was set to `muzzle`, which contained a garbage value from a
random player. For the zap entity, the origin's only purpose is to
determine whether it is in someone's PVS. So the bug caused the zap to
appear and disappear for a given client depending on whether the garbage
value is in its PVS. In a quick test with 5v5, the entity for a single
zap would frequently leave and reenter 5+ times.