Fix bug where bot fails to buy its chosen equipment
The case I saw is that a bot carrying the painsaw decided to purchase
armor instead. It couldn't do it because it did not sell weapons first.

It seems that this code would also be wrong in the case that a bot is
carrying an upgrade and decides it no longer wants it (it wouldn't sell
the upgrade)