Fix some issues with missile tracing
Missiles with pointAgainstWorld did not explode when they were
overlapping a player bounding box but not fully contained with it. This
often happens because because missiles trace against players but not
vice versa, so if the player "moves first" it can be overlapping the
missile. Fix this by using G_Trace2 which always collides with
overlapping entities.

Also ensure that if the missile collides with something, there is always
a valid normal vector. (Tracing doesn't give one in the cases where the
start poing is overlapping.)