Gameplay bug fix for lucifer cannon overcharge
The luci is supposed to explode in place when you overchange it, but
actually it travels a short distance (however far it can go in
MISSILE_PRESTEP_TIME (50) plus one frame) before exploding. So it is
actually possible to hit a very close entity for 265 damage when
overcharging. This seems like a bug to me. Now that it explodes in place
you can only do about 120 damage to a close entity.

You can still luci-jump by overcharging toward the ground same as
before, altough I don't understand why this still works.