Don't set hive missile velocity to 0 on death
The motivation for this was to simplify the code in preparation for
porting to CBSE. However it is also an aesthetic improvement. With the
old behavior, you typically get a cluster of bees spinning in a tight
circle. The spinning, alpha-fading circle of bees looks stupid. With the
new behavior they tend to disperse in various directions which seems
better.