UI: remove the ribbon display on map loading screen
While I think the ribbon is a good idea, it also conflicts with many,
many levelshots which print various information about the map in a much
nicer nicer way.
Not only the ribbon is redundant, but it also hides useful ambiant texts
and can not be disabled.
Only unvanquished official maps and some rare maps without any levelshot
really use that feature, and they're never the most interesting maps.