Fix double-send of info queries on server refresh
When refreshing the serverlist, the "globalservers" or "localservers"
console command is used to make the engine query the master server again
and rebuild the list of potential servers. The command is buffered and
not executed until after the cgame frame finishes. Meanwhile the cgame
immediately sends trap_LAN_ResetPings and trap_LAN_UpdateVisiblePings
which causes pings to be sent for all the servers in the *old* list. As
soon as the new list arrives from the master, everything on the new list
is pinged as well.

Also ensure that the servers are set "visible" as soon as possible so
that pings will be sent out as soon as possible.