README: cd into build for cmake commands
The commands previously are great for scripting due to being
generator-agnostic and not changing the directory, but are not good for
interactive use. The traditional style is less verbose and easier to
remember for interactive use. Plus other sections of the readme tell you
to cd into the build dir to run ./daemon anyway.

Also fixes wrong path for toolchain file.