Fix freezing the mouse when binding a key
When you select an action to bind in the bind menu, the cursor was
supposed to be frozen until a key is chosen or binding is canceled. This
worked on 0.51.1. Probably it was broken by RmlUi upgrades.

The old behavior could be exactly restored by just adding
    rocketInfo.cursor_pos.x = mouse_x;
    rocketInfo.cursor_pos.y = mouse_y;
in the "mousemove" handler, but I decided to rewrite it for a couple of
improvements:
1. Avoid doing ProcessMouseMove inside a mouse event handler. Actually
   it's unclear why this does not cause infinite recursion...
2. Immediately restore the cursor to its "true" position once the
   binding is over. The old behavior is the cursor would jump to its
   true position only the next time the mouse was moved.