Fix escaping from key binding closing the menu
Fixes #2896.

The problem was that the document's event handle for closing when escape
is pressed was receiving the event first, before the key binding
element. Make the key bind element get the event first by registering
the event handle for the "capture" phase, which comes first, instead of
the "bubble" phase.