Let binds be active while using some circlemenus
It is now possible to use binds while the evolve menu, build menu, or
beacon menu is open. So, for example, you can walk forward while
choosing an alien class, or use wallwalk in non-toggling mode while
choosing a structure as Advanced Granger.

A menu may be configured to allow binds at the same time by adding the
"passthrough" keyword in rocket.txt. This means that the menu will only
capture keystrokes that do something in the menu (e.g. numbers in circle
menus), while unconsumed keystrokes will be processed as binds. Note
that mouse clicks and mouse movements are still always captured by any
open menu.

I excluded the armory menu because it might be frustrating for purchases
to fail due to having moved too far from the armory.