Print ladders in command instead of navgen; redo algo
Instead of printing out coordinates of ladders as a side effect of
navgen, print them out when a /printLadders command is issued. This
should be better for people who don't care about ladders (who now won't
receive log spam) and people who do (who can print them quickly without
having to wait for navgen).

Rewrite the algorithm for deciding the ladder endpoints. The old
algorithm seemed to assume that there could only be two unique z
coordinates among the ladder face vertices. The new algorithm gives
the same result for the majority of ladders but
sometimes it is better. I didn't see any where it is worse.

One difference which may or may not be considered worse is that movable
ladders like in Chasm are now included.

Use %.1f instead of %d to print the float values. %d seemingly printed a
random number of digits of precision.