Add r_mode -2 option in the resolution list
r_mode -2 means resolution the same as the active display.

Fixes #1892 (screen resolution not available in the resolution list).

It's nice to have an r_mode -2 option, rather than just one to set the
custom width/height to the current resolution, since this way it
automatically adapts if you connect to different displays.