Bot pathfinding: fix coordinate conversion error
When attempting to detect whether a BOT_TARGET_DYNAMIC target has moved
since the last search, the target's size in the z dimension was used
instead of the intended (I assume) y dimension. The consequences of this
are probably not too serious: it only means that a target with maxs[2]
greater than maxs[0/1] could appear slightly wider than it is.

The PointInPolyExtents function takes an 'rVec extents' argument. But it
appears the function was treating it as a Quake-format vector. One of
the callers of this function was passing a Quake-format vector while the
othew was passing a Recast-format vector.