Fix GLM indexing UB with writable vectors
See https://github.com/Unvanquished/Unvanquished/issues/3013.
To fix the problem in cases where a GLM vector is passed to a function
that wants to write to vec3_t, a proxy object is introduced, which
copies the glm::vec3's content's into a temporary float array, then
copies the gloat array back to the glm::vec3 upon destruction. This
proxy object is spawned by the GLM4RW macro.