locdamge: Load locational damage config based on class
Before, we were loading it based on the model name.

This practically affect anything since this mainly affects aliens, which
do not consider locational damage, but its still a bug and affects the
human-iqm branch since now the human_male model is the same for naked,
light armor, and medium armor classes, which meant that they all used
the same locational damage, which is wrong.