Don't store 'is bot' bit in svFlags
Stop using SVF_BOT and make a bool isBot in gclient_t intsead. SVF_
flags used by the sgame to request certain behaviors from the server.
But SVF_BOT is not used by the server at all (it has its own isBot bool
elsewhere) so it shouldn't be in svFlags.