Fix evolve room-finding when flush with wall
Fixes a regression in 47702c121585f1cd39d8b12a3192b2e4766f0e31 which
made it so that the game can't automatically find room to evolve if you
are exactly against an (axis-aligned) wall.

The room-finding algorithm was moving away from the origin only by the
exact minimum difference between bounding box sizes (thus dubiously
relying on the exactness of a floating-point calculation). I assumed
that if the trace goes 0 distance, it must be stuck and we can assume
failure for the given position. But it turns out the trace often goes 0
distance but is NOT stuck in an obstacle because it was positioned such
that the new bounding box will be flush with the wall if the old one is.

I added 0.5 qu to the distance to avoid this problem. Also the change
from (fraction == 0.0f) to (trace.allsolid) would fix the problem on its
own, but it's better not to rely on exact distances.

Thanks to freem for identifying the blameful commit.