buildables: make g_indestructibleBuildables change existing buildables
Previously, g_indestructibleBuildables only took effect for buildables
set after the value was set. This was annoying to use because it meant
that this value *must* be set before a map is loaded to make it take
effect. Further, if you want to disable it, it means you need to change
the map for it to stop making buildables invincible.

The new behavior makes the cvar take effect immediately, which is less
surprising and more inline with other cvars.