Use a logger for G_LogPrintf messages
Print console messages from G_LogPrintf using a logger. It has a
different default level depending on whether the server is in a client,
to maintain by default the behavior of showing the messages on a ded
server and hiding them for local games (this can be changed via
cvar logs.level.sgame.gameEvents).

Use tinyformat instead of C sprintf/varargs for G_LogPrintf.