Introduce jetpack navcons
Add code for bots to traverse certain navcons with the jetpack.

A navcon can be designated as for jetpack wearers only by setting the
poly flags (3rd-to-last field in a text-format navcon) to 2. When a bot
buys or sells equipment, its pathfinding query filter will be updated so
that it uses or doesn't use the navcon as appropriate.

Note that the poly flags are not used for the bot's small-scale
movement. Jetpack usage is triggered based on z coordinate differences
of the navcon.

Co-Authored-By: slipher <slipher@protonmail.com>