Don't do CBSE generation in the NaCl subproject
When building the NaCl sgame, don't generate the CBSE files in the NaCl
build. Instead depend on the top-level CMake project to generate them
and force a dependency between the nacl-vms target and the generated
files.

This fixes the regression in ce7d25c3224ce4af5c77c3847ce829455e47681b
(rename DAEMON_CBSE_PYTHON_PATH -> CBSE_PYTHON_PATH) that the Python
path was no longer passed to the subproject. Also it makes sure that two
instances of the CBSE generator would not try to run at the same time.