Remove r_halfLambertLighting from presets
In the interest of not having anything which is set to the same value in
all 5 presets. If the value is the same in all presets (and equal to the
cvar's default value), the only thing the preset can do is clobber a
setting which was explicitly chosen by the user.

r_clear can stay since it interacts with r_fastsky which is set by the
presets.