Fix first dropdown selection ignored with custom value
With a cvar-backed <select> element, if the cvar's initial value was not
in the provided list of options, the cvar did not get updated the first
time the user selected an element. The bug was setting
ignore_value_change after SetSelection instead of before. Fixes #3115.

Also remove unnecessary code for filtering out selection of the already
selected element. RmlUi already does this. (And it's wrong IMO, if the
user clicks again the cvar ought to be set again. Oh well.)