Simplify location-dependent color grade code
I verified that the new code gives the same results on plat23
version b11, which is the only map to have ever
used the feature. The color grades there are weak compared to the colors
of the map's lighting, so I used fullbright mode to easily see it.

The new code can give different (and more continuous) results in cases
where there are 3 or more color gradings in range.