emoji: Add emoji support to rmlui.
You can use [:)] or [salute] or other known emoji short codes to use
emojis in Unvanquished. These are a critical feature to help player
express themselves.

This change generates an emoji map using Google's emoji data to map
short names and emoticons to emojis and also provides a script to
regenerate this should a new emoji spec come out.

This also doesn't support combination emojis or emojis using a zwj (so
this means no emoji variations, etc). If we want this support, we'll
need to use the HarfBuzz library to shape the text.