Fix realLight keyword for particle coloring
A 0-1 color needed to be rescaled to 0-255. This regression may have
been caused by Daemon 07251fce, which appears to change the
R_LightForPoint results from a 0-255 scale to a 0-1 scale. But I didn't
verify.

For now, realLight is only really fixed with overbright clamping enabled,
as R_LightForPoint does not scale with the light factor as it should.

Also fix the same issue in trails. Although no one has ever used
realLight in a trail.