Fix BP vampire element drawing when it shouldn't
When BP vampire mode is off, the element still tried to draw a rectangle
(with NaN coordinates). Make sure to check that the mode is on before
doing anything. This caused an out-of-bounds vector access in the
material system for me (when combined with renderer bugs).

Also split updating/rendering into OnUpdate/OnRender.