Vampire mode: re-enable the BP queue, free BP as in mining
Fix this bug: when a building under construction is destroyed by the enemy team, the used and the free BP of the losing team do not add up to the total BP anymore. This commit requires the BP recovery rate to be below about 90 per minute, that is 1.5 BP per second or 6 BP in 4 seconds. The time between death and explosion is about 4 seconds. BP recovery starts at death, but the vampire BP transfer is done at explosion. The team must not recover enough BP to rebuild in this time.