Vampire mode: fix a bug related to the marked BP budget
To see the bug, spend all your BP. Then mark a building, let the enemy team damage it, and in the next 5 seconds, replace it by a new building. After the dust has settled, your team is using more BP than you have.

The reason is that we subtract half of the building's BP when it is replaced and was damaged recently. This commit makes the marked BP counting aware of that.