presets: disable bloom for now
Disable bloom because it is not usable with none of our current maps.

It was enabled by mistake because it didn't looked bad while the light was
mistakenly clamped. The buggy clamping prevented bloom to happen most of the
time. A bad computation was breaking another bad computation and they were
cancelling themselves together.

The future of bloom is either:

- We NUKE bloom entirely.
- We enable bloom on maps having lights computed in linear space,
  and we disable it on maps having lights not computed in non-linear space.
- We enable bloom on maps having lights computed in linear space,
  and we enable it on maps having lights not computed in non-linear space
  BUT with a different set of default values specially crafted for such maps.

The current bloom option and values have no future with all maps that have
been already published and that will be published before maps use lights
computed in linear space.