Fix framerate-dependent animation blending
The rate of transition between animations depended not only on the
cg_animBlend, but also on the frame rate. Given a frame length in ms of
at least 11 ms and a value of the old cvar cg_animBlend, you could
achieve the same transition rate by setting the new cvar
cg_animBlendHalfLife as follows:
cg_animBlendHalfLife = -frametime / log2(1 - 1/cg_animBlend)

So for the old default of cg_animBlend = 5, a 100 ms frametime (slow
blending) equates to cg_animBlendHalfLife = 310.6. For 11 ms frametime
(fastest blending), that's 34.17. For frame times less than 11 ms (~91
FPS), the old algorithm didn't advance the blend fraction every frame so
the half life would wobble a bit above the minimal value.

A good place to observe the effects is with the barricade raise/lower
animations.  This is because they are poorly made s.t. the last frames
don't align well with the following idle animations.