Fix keynum_t binds in bind menu
A regression caused bindable key events not to be sent for any
Keyboard::Key that can be translated to an RmlUi key. This is because
the bindable key event started being disabled if the normal key event
was consumed, but the normal key event was always consumed. This just
affected keynum_t Key's since character ones are ignored anyway.

Instead of sending both normal RmlUi key events and bindable key events
all the time, send only bindable key events during key binding and only
normal key events otherwise. This should be a more reliable design
anyway since some keys such as Tab and Enter could be interpreted as
ones for navigating the document. So now Escape for canceling a bind
will be detected through the bindable key event instead of the normal
key event.

Fixes #3454.