UI: improve positioning of "Beta x.y.z." label
The positioning of the version label on the in-game menu looked bad.
Make sure it is at the bottom left. CSS for the main menu label is also
reworked. The two menus needed some different methods for vertical
positioning since in the main menu the label needs to allocate some
vertical space, while in the ingame menu I wanted to use the space
already allocated for the quit button.